Add interface for table loader
